rename Certifications component to Expertise
The section now displays technical expertise areas rather than certifications,
so the component name should reflect its actual content.
