+++ src/lib/components/sections/Expertise.svelte
